update client kick

+++ ygopro-server.coffee
@@ -441,7 +441,10 @@ ROOM_unwelcome = (room, bad_player, reason)->
 
 CLIENT_kick = (client) ->
   client.system_kicked = true
-  client.destroy()
+  if settings.module.reconnect.enabled and client.closed
+    CLIENT_reconnect_unregister(client, false, true)
+  else
+    client.destroy()
   return
 
 release_disconnect = (dinfo, reconnected) ->
